user - audiencesurvey page

+++ app/api/authentication/logout/route.ts
@@ -1,6 +1,7 @@
 import { NextResponse } from "next/server";
 import { APIRequestOptions, fetchData } from "../../FetchData";
-import { getToken, removeToken } from "../../Session";
+// import { getToken, removeToken } from "../../Session";
+import { cookies } from "next/headers";
 
 interface ApiResponse {
   data?: any;
@@ -10,10 +11,16 @@ interface ApiResponse {
 
 export async function DELETE(req: Request) {
   const backendUrl = process.env.NEXT_PUBLIC_BACKEND_URL;
-  const body = await req.json();
-  const token = body?.token ?? null;
 
   const path = `${backendUrl}/auth`;
+  const cookie = (await cookies()).get("authToken")?.value;
+
+  if (!cookie) {
+    throw new Error("missing auth token");
+  }
+
+  const parsedCookie = JSON.parse(cookie) as { token: string; role?: string };
+  const token = parsedCookie.token;
   const options: APIRequestOptions = {
     method: "DELETE",
     headers: {
@@ -23,8 +30,8 @@ export async function DELETE(req: Request) {
   };
   try {
     const response = await fetchData<ApiResponse>(path, options);
-    const res = NextResponse.json(response);
-    removeToken(res, "authToken");
+    var res = NextResponse.json<ApiResponse>(response);
+    (await cookies()).delete("authToken");
     return res;
   } catch (error) {
     return NextResponse.json({
